feat(asimov): SONIC embodiment port — verified assets, registration, tooling
Asimov v1 (23 DOF) as a SONIC training target, mirroring the X2 port with
generalizations and hard-won asset verification:

- MJCF staged from mjlab f5c4caa; welded neck fused into torso_link
  (bodies==1+23), framework body renames, actuator block; fusion composes
  RAW XML transforms (compiled geom_quat embeds mesh alignment — double-
  apply detached the head), verified: 48/48 geoms at 4e-10, mass+COM exact
- URDF generated with FK-at-zero (elbow ref baked, parity 7nm/200 poses),
  collision = the MJCF's real 21 primitives (capsules+foot spheres; mesh
  colliders poisoned run 1 with phantom self-collisions), visuals baked
  into per-link STLs from compiled mesh_vert (importer drops nonzero
  visual origins). Visually verified in MuJoCo AND IsaacLab.
- robots/asimov.py: hardware-characterized gains (asimov_1_constant),
  mirrored per-side default pose; mapping_utils generates IL<->MJ maps
  from name lists (X2-regression-tested); runtime joint+body order
  verified == prediction (gate C1)
- registered: robot_mapping, AsimovConverter, --robot asimov pkl builder
  (float joint axes for the canted wrist), sonic_asimov_loco exp yaml
  (wrist joints_idx override for 23 DOF)
- eval/tooling: exec_qpos capture in im_eval dump, MuJoCo overlay viewer
  (solid=executed vs ghost=reference), Isaac kinematic replay gate,
  order-check gate, ckpt watcher; asimov_eval8.pkl held-out set

Lesson encoded in the dev scripts: frame-level numeric gates do not
verify geom placement — render-and-look is a mandatory gate after any
asset transformation.

+++ gear_sonic/utils/motion_lib/torch_humanoid_batch.py
@@ -196,18 +196,18 @@ class Humanoid_Batch:
             and tree.getroot().find("worldbody").findall(".//joint")[0].attrib["type"] == "free"
         ):
             for j in tree.getroot().find("worldbody").findall(".//joint")[1:]:
-                self.dof_axis.append([int(i) for i in j.attrib["axis"].split(" ")])
+                self.dof_axis.append([float(i) for i in j.attrib["axis"].split(" ")])  # float: canted axes (Asimov wrist) are valid
             self.has_freejoint = True
         elif "type" not in tree.getroot().find("worldbody").findall(".//joint")[0].attrib:
             for j in tree.getroot().find("worldbody").findall(".//joint"):
-                self.dof_axis.append([int(i) for i in j.attrib["axis"].split(" ")])
+                self.dof_axis.append([float(i) for i in j.attrib["axis"].split(" ")])  # float: canted axes (Asimov wrist) are valid
             self.has_freejoint = True
         else:
             for j in tree.getroot().find("worldbody").findall(".//joint")[6:]:
-                self.dof_axis.append([int(i) for i in j.attrib["axis"].split(" ")])
+                self.dof_axis.append([float(i) for i in j.attrib["axis"].split(" ")])  # float: canted axes (Asimov wrist) are valid
             self.has_freejoint = False
 
-        self.dof_axis = torch.tensor(self.dof_axis)
+        self.dof_axis = torch.tensor(self.dof_axis, dtype=torch.float32)
 
         for extend_config in cfg.extend_config:
             self.body_names_augment += [extend_config.joint_name]
